Added new way to read in corpus.
The new way of reading expects a corpus to be all in one file, where each line
represents a document.  The first non-whitespace characters of the line are the
title for a given document, and the rest of the line (separated from the title
non-newline whitespace) is composed of the document's words.